New tracklist based on tracksForDisc in HMV
